requirement.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
Django==4.2.11
django-crispy-forms==2.1
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
idna==3.7
MarkupSafe==2.1.5
numpy==2.0.0rc1
pandas==2.2.2
ply==3.11
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rfc3986==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
thriftpy2==0.4.20
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
